all controllers currently working

+++ models/moviesDB.sql
@@ -1,10 +1,10 @@
 CREATE DATABASE moviesDB;
 USE moviesDB;
-CREATE TABLE movies(
+CREATE TABLE movies (
 	id INT AUTO_INCREMENT NOT NULL,
     movieTitle VARCHAR(50) NOT NULL,
     movieYear INTEGER,
-    movieRating INTEGER,
+    movieRating DECIMAL(1, 2),
     movieGenre VARCHAR(50),
     movieComments VARCHAR(50),
     movieSad BOOLEAN,
